Home Portfolio card issue fixed

+++ networklayer/src/main/java/com/emproto/networklayer/response/home/PortFolioData.kt
@@ -2,7 +2,7 @@ package com.emproto.networklayer.response.home
 
 data class PortFolioData(
     val amountInvested: Double,
-    val amountPending: Int,
+    val amountPending: Double,
     val investmentCount: Int,
-    val totalAreaSqFt: Int
+    val totalAreaSqFt: Double
 )
